+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade `@jupiterone/integration-sdk-\*@5.3.0
+- Adopted `indexMetadata` in each step's graph object metadata to omit disk
+  writes
+
 ## 4.7.5 - 2020-12-22
 
 ### Changed